requirements.txt
python-dotenv
requests

langchain-community
langchain-core
langchain
langchain-huggingface
langchain-google-genai

chromadb